fixed clientside and server side issues with metadata
